Improvements in wireless code: - use all available pipes to connect to motor addresses and use those channels to send and receive messages instead of using the remote controll addres for both - some optimizations for wirelessMessage struct (use byte instead of int where possible) - don't wait for next loop to check for ack message, but check for radio.available() in the same execution thread until there is data available (this blocks the current loop) - don't send array of all addresses when sending message to all motors, use a new method and use indexes to decide which address to send to - better logging

+++ moving-sculptures/modules/motor.h
@@ -11,9 +11,9 @@ struct motorParameters {
 
 struct motorStatus {
 	int speed;
-	int direction;
-	int cs1;
-	int cs2;
+	byte direction;
+	byte cs1;
+	byte cs2;
 	int battery;
 	bool disabled;
 };
